Optimize the index building. (#5)
We skip the sort during build if subjects are already sorted. This is a useful
optimization for many common use cases.g., sorted BAM files, bigWigs.

We also avoid allocating many small vectors in the working list, which is
expensive to create, access and deallocate. Instead, we create one big vector
where the right-hand-side holds the children for the currently active nodes and
the left-hand-side holds the children for already-processed nodes. Then we
simply move children from the left to the right once a node is processed.